+++ Classes/HtmlHelpers.cs
@@ -33,24 +33,24 @@ namespace VidsNet.Classes
 
         public static TagBuilder GenerateItem(string type, List<string> cssClasses, Dictionary<string, string> attributes, 
         TagBuilder innerData = null, string innerText = null) {
-            var span = new TagBuilder(type);
+            var item = new TagBuilder(type);
 
             foreach (var cssClass in cssClasses)
             {
-                span.AddCssClass(cssClass);
+                item.AddCssClass(cssClass);
             }
 
-            span.MergeAttributes(attributes);
+            item.MergeAttributes(attributes);
 
             if(innerText != null) {
-                span.InnerHtml.Append(innerText);
+                item.InnerHtml.Append(innerText);
             }
 
             if(innerData != null) {
-                span.InnerHtml.AppendHtml(innerData);
+                item.InnerHtml.AppendHtml(innerData);
             }
 
-            return span;
+            return item;
         }
 
         public static void GenerateDirectoryListing(ScanItem item, TagBuilder select, int level, List<UserSetting> userPaths) {
@@ -64,29 +64,14 @@ namespace VidsNet.Classes
                 var checkbox = HtmlHelpers.GenerateCheckbox(Path.GetFileName(child.Path), pathValue, child.Id.ToString(), isChecked);
 
                 var div = HtmlHelpers.GenerateItem("div", new List<string>() { "list-group-item"}, new Dictionary<string, string>() {
-                    { "style", string.Format("padding-left: {0}px", padding) } }, checkbox);
+                    { "style", string.Format("padding-left: {0}px", padding) }, { "id", string.Format("{0}_div", item.Id)} }, checkbox);
 
                 if(child.Children.Count > 0) {
-                    /*var nameSpan = HtmlHelpers.GenerateItem("span", new List<string>(), new Dictionary<string, string>() { 
-                        { "id", string.Format("{0}_name", guid)  }, { "style", "margin-left: 15px;"} }, null, Path.GetFileName(child.Path));
-
-                    var buttonSpan = HtmlHelpers.GenerateItem("span", new List<string>() { "glyphicon","glyphicon-folder-open"}, 
-                    new Dictionary<string, string>() { { "aria-hidden", "true" } }, nameSpan);
-
-                    var button = HtmlHelpers.GenerateItem("button", new List<string>() { "btn", "btn-link" }, new Dictionary<string, string>() { 
-                        { "data-toggle", "collapse" }, { "data-target", string.Format("#{0}", guid) }, { "type", "button" } }, buttonSpan);*/
                     var button = GenerateFolder(Path.GetFileName(child.Path), guid, true);
-
                     div.InnerHtml.AppendHtml(button);
                 }
                 else {
-                    /*var nameSpan = HtmlHelpers.GenerateItem("span", new List<string>(), new Dictionary<string, string>() { 
-                        {"id",string.Format("{0}_name", guid)  }, {"style", "margin-left: 15px;"} }, null, Path.GetFileName(child.Path));
-
-                    var buttonSpan = HtmlHelpers.GenerateItem("span", new List<string>() { "glyphicon","glyphicon-folder-open"}, 
-                    new Dictionary<string, string>() { { "aria-hidden", "true"}, { "style", "padding: 6px 12px" } }, nameSpan);*/
                     var button = GenerateFolder(Path.GetFileName(child.Path), guid, false);
-
                     div.InnerHtml.AppendHtml(button);
                 }
 
@@ -116,7 +101,7 @@ namespace VidsNet.Classes
                     new Dictionary<string, string>() { { "aria-hidden", "true" } }, nameSpan);
 
                 var button = HtmlHelpers.GenerateItem("button", new List<string>() { "btn", "btn-link" }, new Dictionary<string, string>() { 
-                { "data-toggle", "collapse" }, { "data-target", string.Format("#{0}", guid) }, { "type", "button" } }, buttonSpan);
+                { "data-toggle", "collapse" }, { "data-target", string.Format("#{0}_content", guid) }, { "type", "button" } }, buttonSpan);
 
                 return button;
 
@@ -134,18 +119,29 @@ namespace VidsNet.Classes
             
         }
 
-        public static TagBuilder GenerateViewFolder(BaseVirtualItem item, int padding, int childCount) {
+        public static TagBuilder GenerateViewFolder(BaseVirtualItem item, int padding, int childCount, View viewType, int? parentId = null) {
+
             var div = GenerateItem("div", new List<string>() { "list-group-item"}, new Dictionary<string, string>() {
-                { "style", string.Format("padding-left: {0}px", padding) } });
+                { "style", string.Format("padding-left: {0}px", padding) }, { "id", string.Format("{0}_div", item.Id)} });
             
             var hasItems = childCount > 0;
             var button = GenerateFolder(Path.GetFileName(item.Name), item.Id.ToString(), hasItems);
 
             var count = GenerateItem("span", new List<string>() { "label", "label-default", "label-pill", "pull-xs-right"},
-             new Dictionary<string, string>(), null, childCount.ToString());
+             new Dictionary<string, string>() { {"id", string.Format("{0}_count", item.Id.ToString())} }, null, childCount.ToString());
 
+            var optionsSpan = GenerateItem("span", new List<string>(){ "glyphicon", "glyphicon-menu-hamburger" }, new Dictionary<string, string>() {
+                { "aria-hidden", "true" } });
+
+
+            var optionsButton = HtmlHelpers.GenerateItem("button", new List<string>() { "btn", "btn-link" }, new Dictionary<string, string>() { 
+                { "data-toggle", "popover" }, {"data-placement", "top"}, {"data-content", TagToString(GenerateViewItemOptions(item, viewType, parentId)) },
+                 {"id", string.Format("viewItem_{0}", item.Id)},  { "type", "button" }, 
+                 {"tabindex", item.Id.ToString() } }, optionsSpan);
+            
             div.InnerHtml.AppendHtml(button);
             div.InnerHtml.AppendHtml(count);
+            div.InnerHtml.AppendHtml(optionsButton);
 
             return div;
         }
@@ -158,7 +154,7 @@ namespace VidsNet.Classes
             var button = GenerateFolder(Path.GetFileName(item.Name), item.Id.ToString(), hasItems);
 
             var count = GenerateItem("span", new List<string>() { "label", "label-default", "label-pill", "pull-xs-right"},
-             new Dictionary<string, string>(), null, childCount.ToString());
+             new Dictionary<string, string>() , null, childCount.ToString());
 
             div.InnerHtml.AppendHtml(button);
             div.InnerHtml.AppendHtml(count);
@@ -166,8 +162,65 @@ namespace VidsNet.Classes
             return div;
         }
 
-        public static TagBuilder GenerateViewItem(BaseVirtualItem item, int padding, string url) {
-            var div = GenerateItem("div", new List<string>() { "list-group-item"}, new Dictionary<string, string>());
+        public static TagBuilder GenerateViewItemOptions(BaseVirtualItem item, View viewType = View.Default, int? parentId = null) {
+            var div = new TagBuilder("div");
+            var padding = string.Format("padding: {0}px {1}px;", 0,5);
+            if(!parentId.HasValue) {
+                parentId = item.ParentId;
+            }
+            if(viewType == View.Viewed) {
+                var viewed = GenerateItem("span", new List<string>() {"glyphicon", "glyphicon-eye-close", "icon-small" }, new Dictionary<string, string>() {
+                { "aria-hidden", "true" }, {"style", padding} });
+                var viewedLink = GenerateItem("button", new List<string>() { "btn", "btn-link"}, new Dictionary<string, string>() {
+                    {"onclick", string.Format("unviewedItem({0}, {1});", item.Id, parentId)} }, viewed);
+                div.InnerHtml.AppendHtml(viewedLink);
+            }
+            else if(viewType == View.Deleted) {
+                var delete = GenerateItem("span", new List<string>() {"glyphicon", "glyphicon-ok", "icon-small" }, new Dictionary<string, string>() {
+                    { "aria-hidden", "true" }, {"style", padding} });
+                var deleteLink = GenerateItem("button", new List<string>() { "btn", "btn-link"}, new Dictionary<string, string>() {
+                    {"onclick", string.Format("undeleteItem({0}, {1});", item.Id, parentId)} }, delete);
+                div.InnerHtml.AppendHtml(deleteLink);
+            }
+            else {
+                var viewed = GenerateItem("span", new List<string>() {"glyphicon", "glyphicon-eye-open", "icon-small" }, new Dictionary<string, string>() {
+                { "aria-hidden", "true" }, {"style", padding} });
+                var viewedLink = GenerateItem("button", new List<string>() { "btn", "btn-link"}, new Dictionary<string, string>() {
+                    {"onclick", string.Format("viewedItem({0}, {1});", item.Id, parentId)} }, viewed);
+
+                
+                var edit = GenerateItem("span", new List<string>() {"glyphicon", "glyphicon-pencil", "icon-small" }, new Dictionary<string, string>() {
+                    { "aria-hidden", "true" }, {"style", padding} });
+                var editLink = GenerateItem("button", new List<string>() { "btn", "btn-link"}, new Dictionary<string, string>() {
+                    {"onclick", "editItem("+item.Id+");"} }, edit);
+                
+                
+                var delete = GenerateItem("span", new List<string>() {"glyphicon", "glyphicon-remove", "icon-small" }, new Dictionary<string, string>() {
+                    { "aria-hidden", "true" }, {"style", padding} });
+                var deleteLink = GenerateItem("button", new List<string>() { "btn", "btn-link"}, new Dictionary<string, string>() {
+                    {"onclick", string.Format("deleteItem({0}, {1});", item.Id, parentId)} }, delete);
+
+                var move = GenerateItem("span", new List<string>() {"glyphicon", "glyphicon-move", "icon-small" }, new Dictionary<string, string>() {
+                    { "aria-hidden", "true" }, {"style", padding} });
+                var moveLink = GenerateItem("button", new List<string>() { "btn", "btn-link"}, new Dictionary<string, string>() {
+                    {"onclick", "moveItem("+item.Id+");"} }, move);
+
+                div.InnerHtml.AppendHtml(viewedLink);
+                div.InnerHtml.AppendHtml(editLink);
+                div.InnerHtml.AppendHtml(moveLink);
+                div.InnerHtml.AppendHtml(deleteLink);
+            }
+
+            
+
+            
+
+            return div;
+        }
+
+        public static TagBuilder GenerateViewItem(BaseVirtualItem item, int padding, string url, View viewType = View.Default, int? parentId = null) {
+            var div = GenerateItem("div", new List<string>() { "list-group-item"}, new Dictionary<string, string>() {
+                 { "id", string.Format("{0}_div", item.Id)} });
             
             var nameSpan = GenerateItem("span", new List<string>(), new Dictionary<string, string>() {
                 { "id", string.Format("{0}_name", item.Id) } }, null, item.Name);
@@ -175,15 +228,25 @@ namespace VidsNet.Classes
                 { "href", url }, { "style", "margin-left: 15px;" } }, nameSpan);
 
             var iconSpan = GenerateItem("span", new List<string>(){ "glyphicon", "glyphicon-film"}, new Dictionary<string, string>() {
-                { "aria-hidden", "true" }, { "style", string.Format("padding-left: {0}px", padding) } }, nameLink);
+                { "aria-hidden", "true" }, { "style", string.Format("padding-left: {0}px;", padding) } }, nameLink);
 
+            var optionsSpan = GenerateItem("span", new List<string>(){ "glyphicon", "glyphicon-menu-hamburger" }, new Dictionary<string, string>() {
+                { "aria-hidden", "true" } });
+
+            var button = HtmlHelpers.GenerateItem("button", new List<string>() { "btn", "btn-link" }, new Dictionary<string, string>() { 
+                { "data-toggle", "popover" }, {"data-placement", "top"}, {"data-content", TagToString(GenerateViewItemOptions(item, viewType, parentId))},
+                 {"id", string.Format("viewItem_{0}", item.Id)},  { "type", "button" },
+                 {"tabindex", item.Id.ToString() } }, optionsSpan);
+            
             div.InnerHtml.AppendHtml(iconSpan);
+            div.InnerHtml.AppendHtml(button);
             return div;
 
         }
 
         public static TagBuilder GenerateViewItem(RealItem item, int padding, string url) {
-            var div = GenerateItem("div", new List<string>() { "list-group-item"}, new Dictionary<string, string>());
+            var div = GenerateItem("div", new List<string>() { "list-group-item"}, new Dictionary<string, string>() {
+                 { "id", string.Format("{0}_div", item.Id)} });
             
             var nameSpan = GenerateItem("span", new List<string>(), new Dictionary<string, string>() {
                 { "id", string.Format("{0}_name", item.Id) } }, null, item.Name);