+++ DataModels/MoveItem.cs
@@ -1,0 +1,6 @@
+namespace VidsNet.DataModels
+{
+    public class MoveItem {
+        public int ParentId {get;set;}
+    }
+}