+++ Enums/View.cs
@@ -1,0 +1,8 @@
+namespace VidsNet.Enums
+{
+    public enum View {
+        Default,
+        Viewed,
+        Deleted
+    }
+}